Rewrite external url rewrite to a hook
